Re-write how function hijacking is done
This commit ditches the method of writing an asm jump code to the start of a
function to hijack it, and instead replaces the pointer to that function in the
appropriate place. I pretty much lucked out here - all of the functions I end
up needing to hijack are called from an operations structure of one sort or
another, otherwise this switcharoo wouldn't be possible.

I dare say this method is "safer", but this is the initial work so I don't dare
lable it "stable" yet :) So far I've only tested it on linux 3.2.0, though I'm
fairly certain this will back-port just fine to 2.6.32 and even 2.6.18. I'm
sure there are going to be a few bugs to work out, this is the "dev" branch
after all :)

The only regression I'm aware of at this point is I'm not currently checking if
the functions are already hijacked, and I'm not even sure if that's really
possible with this method. Oh and, the code is somewhat messier in places,
cleaner in others.

An added bonus of this change - this module should now work on linux systems
that aren't on a x86 CPU set. I don't have any of those kinds of systems, but
I'll ask around for testers.